+++ src/Commodity.java
@@ -1,6 +1,6 @@
 public interface Commodity {
     int getCommodityNum();
 
-    int getCommodityValue();
+    long getCommodityValue();
 
 }
