+++ src/RecoverBottle.java
@@ -7,7 +7,7 @@
 public class RecoverBottle extends Bottle {
     private double ratio;
 
-    public RecoverBottle(int id, String name, int capacity, int price, double ratio) {
+    public RecoverBottle(int id, String name, int capacity, long price, double ratio) {
         super(id, name, capacity, price);
         this.ratio = ratio;
     }
