+++ src/RegularBottle.java
@@ -5,7 +5,7 @@
  * @author hugo
  */
 public class RegularBottle extends Bottle {
-    public RegularBottle(int id, String name, int capacity, int price) {
+    public RegularBottle(int id, String name, int capacity, long price) {
         super(id, name, capacity, price);
     }
