+++ src/RegularEquipment.java
@@ -6,7 +6,7 @@
  */
 public class RegularEquipment extends Equipment {
 
-    public RegularEquipment(int id, String name, int star, int price) {
+    public RegularEquipment(int id, String name, int star, long price) {
         super(id, name, star, price);
     }
